+++ scripts/install.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-VERSION=2.0.7
+VERSION=2.0.8
 KOTLIN_VERSION=2.1.10
 
 while getopts v:h: flag